backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==1.4
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
asyncpg
databases
python-multipart
aiofiles
aiortc
python-dotenv
firebase-admin